+++ src/sdk/PnP.Core.Test/SharePoint/FilesTests.cs
@@ -1648,6 +1648,70 @@ namespace PnP.Core.Test.SharePoint
             await TestAssets.CleanupTestDocumentAsync(3, contextConfig: TestCommon.NoGroupTestSite);
         }
 
+        [TestMethod]
+        public async Task CopyFileCrossSiteNoOverWriteTest()
+        {
+            //TestCommon.Instance.Mocking = false;
+
+            (_, string documentName, string documentUrl) = await TestAssets.CreateTestDocumentAsync(0);
+
+            using (var context = await TestCommon.Instance.GetContextAsync(TestCommon.TestSite, 1))
+            {
+                using (var otherSiteContext = await TestCommon.Instance.GetContextAsync(TestCommon.NoGroupTestSite, 2))
+                {
+                    IFile testDocument = await context.Web.GetFileByServerRelativeUrlAsync(documentUrl);
+
+                    string destinationServerRelativeUrl = $"{otherSiteContext.Uri.PathAndQuery}/Shared Documents/{documentName}";
+                    
+                    // Perform first copy
+                    await testDocument.CopyToAsync(destinationServerRelativeUrl, options: new MoveCopyOptions { KeepBoth = true });
+
+                    // Perform second copy, this will not overwrite the original file but add a copy with a number suffix
+                    await testDocument.CopyToAsync(destinationServerRelativeUrl, options: new MoveCopyOptions { KeepBoth = true });
+
+                    IFile foundCopiedDocument = await otherSiteContext.Web.GetFileByServerRelativeUrlAsync(destinationServerRelativeUrl);
+                    Assert.IsNotNull(foundCopiedDocument);
+                    Assert.AreEqual(documentName, foundCopiedDocument.Name);
+                }
+            }
+
+            await TestAssets.CleanupTestDocumentAsync(3);
+            await TestAssets.CleanupTestDocumentAsync(3, contextConfig: TestCommon.NoGroupTestSite);
+            // Delete the extra file since we used keepboth twice
+            await TestAssets.CleanupTestDocumentAsync(3, contextConfig: TestCommon.NoGroupTestSite, fileName: $"{Path.GetFileNameWithoutExtension(documentName)}1.docx");
+        }
+
+        [TestMethod]
+        public async Task CopyFileNoOverWriteTest()
+        {
+            //TestCommon.Instance.Mocking = false;
+
+            (_, string documentName, string documentUrl) = await TestAssets.CreateTestDocumentAsync(0);
+
+            string originalFileNameWithoutExt = Path.GetFileNameWithoutExtension(documentName);
+            string copyFileName = $"{originalFileNameWithoutExt}_COPY{Path.GetExtension(documentName)}";
+            string destinationServerRelativeUrl = documentUrl.Replace(documentName, copyFileName);
+
+            using (var context = await TestCommon.Instance.GetContextAsync(TestCommon.TestSite, 1))
+            {
+                IFile testDocument = await context.Web.GetFileByServerRelativeUrlAsync(documentUrl);
+
+                // Perform first copy
+                await testDocument.CopyToAsync(destinationServerRelativeUrl, options: new MoveCopyOptions { KeepBoth = true });
+
+                // Perform second copy, this will not overwrite the original file but add a copy with a number suffix
+                await testDocument.CopyToAsync(destinationServerRelativeUrl, options: new MoveCopyOptions { KeepBoth = true });
+
+                IFile foundCopiedDocument = await context.Web.GetFileByServerRelativeUrlAsync(destinationServerRelativeUrl);
+                Assert.IsNotNull(foundCopiedDocument);
+                Assert.AreEqual(copyFileName, foundCopiedDocument.Name);
+            }
+
+            await TestAssets.CleanupTestDocumentAsync(2, fileName: documentName);
+            await TestAssets.CleanupTestDocumentAsync(2, fileName: copyFileName);
+            await TestAssets.CleanupTestDocumentAsync(2, fileName: $"{Path.GetFileNameWithoutExtension(copyFileName)}1.docx");
+        }
+
         #endregion
 
         #region MoveTo() variants
@@ -3433,5 +3497,40 @@ namespace PnP.Core.Test.SharePoint
         }
 
         #endregion
+
+        [TestMethod]
+        public async Task RenameFileTest()
+        {
+            //TestCommon.Instance.Mocking = false;
+            using (var context = await TestCommon.Instance.GetContextAsync(TestCommon.TestSite))
+            {
+                IFile addedFile = null;
+                try
+                {
+                    IFolder parentFolder = await context.Web.Folders.FirstOrDefaultAsync(f => f.Name == "SiteAssets");
+
+                    string fileName = TestCommon.GetPnPSdkTestAssetName("test_added.docx");
+                    addedFile = await parentFolder.Files.AddAsync(fileName, System.IO.File.OpenRead($".{Path.DirectorySeparatorChar}TestAssets{Path.DirectorySeparatorChar}test.docx"));
+
+                    // Test the created object
+                    Assert.IsNotNull(addedFile);
+                    Assert.AreEqual(fileName, addedFile.Name);
+                    Assert.AreNotEqual(default, addedFile.UniqueId);
+
+                    // Rename file
+                    addedFile.Rename("rename.docx");
+
+                    // Get the file again
+                    IFile foundDocument = await context.Web.GetFileByServerRelativeUrlAsync($"{parentFolder.ServerRelativeUrl}/rename.docx");
+
+                    Assert.IsTrue(foundDocument != null);
+                }
+                finally
+                {
+                    // Cleanup the added file
+                    await addedFile.DeleteAsync();
+                }
+            }
+        }
     }
 }