+++ src/sdk/PnP.Core/Model/SharePoint/Core/Public/IWeb.cs
@@ -892,6 +892,18 @@ namespace PnP.Core.Model.SharePoint
         #region Users
 
         /// <summary>
+        /// Retrieves everyone except external users and ensures the user in the current web
+        /// </summary>
+        /// <returns>The ensured <see cref="ISharePointUser"/></returns>
+        public Task<ISharePointUser> EnsureEveryoneExceptExternalUsersAsync();
+
+        /// <summary>
+        /// Retrieves everyone except external users and ensures the user in the current web
+        /// </summary>
+        /// <returns>The ensured <see cref="ISharePointUser"/></returns>
+        public ISharePointUser EnsureEveryoneExceptExternalUsers();
+
+        /// <summary>
         /// Ensures the given users exists
         /// </summary>
         /// <param name="userPrincipalName">UserPrincipalName value of the user to verify</param>